Добавить таблицу price

+++ src/Controller/ProductController.php
@@ -78,8 +78,13 @@ final class ProductController extends AbstractController
     public function delete(Request $request, Product $product, EntityManagerInterface $entityManager): Response
     {
         if ($this->isCsrfTokenValid('delete'.$product->getId(), $request->getPayload()->getString('_token'))) {
-            $entityManager->remove($product);
-            $entityManager->flush();
+            if ($product->getOrderItems()->count() === 0 ) {
+                $entityManager->remove($product);
+                $entityManager->flush();
+            } else {
+                $this->addFlash('error', 'Данный продукт указан в заказе!');
+            }
+
         }
 
         return $this->redirectToRoute('app_product_index', [], Response::HTTP_SEE_OTHER);
